requirements.txt
nicegui>=1.4.0
httpx>=0.27.0
markdown2>=2.4.0
openai>=1.14.0
pandas>=2.0.0
python-dotenv>=1.0.0
sqlparse>=0.5.0
openpyxl>=3.1.0
fpdf2>=2.7.0
duckdb>=0.10.0
langgraph>=1.0.0
